fix the returned value of the image from preprocessing

+++ supervised_learning/0x0C-neural_style_transfer/0-neural_style.py
@@ -89,4 +89,4 @@ class NST:
         )
         image = image / 255
         image = tf.clip_by_value(image, 0, 1)
-        return image
+        return tf.cast(image, tf.float32)
